+++ refinedstorage-fabric/src/main/java/com/refinedmods/refinedstorage/fabric/ModInitializerImpl.java
@@ -828,20 +828,22 @@ public class ModInitializerImpl extends AbstractModInitializer implements ModIni
 
     private void registerEnergyItemProviders() {
         EnergyStorage.ITEM.registerForItems(
-            (stack, context) -> new EnergyStorageAdapter(Items.INSTANCE.getWirelessGrid().createEnergyStorage(stack)),
+            (stack, context) ->
+                new EnergyStorageAdapter(Items.INSTANCE.getWirelessGrid().createEnergyStorage(stack), context),
             Items.INSTANCE.getWirelessGrid()
         );
         Items.INSTANCE.getControllers().forEach(controller -> EnergyStorage.ITEM.registerForItems(
-            (stack, context) -> new EnergyStorageAdapter(controller.get().createEnergyStorage(stack)),
+            (stack, context) -> new EnergyStorageAdapter(controller.get().createEnergyStorage(stack), context),
             controller.get()
         ));
         EnergyStorage.ITEM.registerForItems(
-            (stack, context) -> new EnergyStorageAdapter(PortableGridBlockItem.createEnergyStorage(stack)),
+            (stack, context) -> new EnergyStorageAdapter(PortableGridBlockItem.createEnergyStorage(stack), context),
             Items.INSTANCE.getPortableGrid()
         );
         EnergyStorage.ITEM.registerForItems(
             (stack, context) ->
-                new EnergyStorageAdapter(Items.INSTANCE.getWirelessAutocraftingMonitor().createEnergyStorage(stack)),
+                new EnergyStorageAdapter(Items.INSTANCE.getWirelessAutocraftingMonitor().createEnergyStorage(stack),
+                    context),
             Items.INSTANCE.getWirelessAutocraftingMonitor()
         );
     }